Movement Script - Gamer
Links the movement of the sprite to the 3D agent that is situated
elsewhere in the environment. Much of the script sets about describing
the different modes in which the sprite can move in, allowing it to
operate under control of the animator and then switch back to following
the agent.